bug sub

+++ src/main/java/com/tms/service/FavoriteServiceCargo.java
@@ -1,12 +1,14 @@
 package com.tms.service;
 
 
-import com.tms.domain.Cargo;
-import com.tms.domain.FavoritesCargo;
+import com.tms.domain.*;
 import com.tms.repository.CargoRepository;
 import com.tms.repository.FavoriteRepositoryCargo;
+import com.tms.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,31 +18,62 @@ public class FavoriteServiceCargo {
 
     FavoriteRepositoryCargo favoriteRepositoryCargo;
     CargoRepository cargoRepository;
+    UserRepository userRepository;
 
 
     @Autowired
-    public FavoriteServiceCargo(FavoriteRepositoryCargo favoriteRepositoryCargo, CargoRepository cargoRepository) {
+    public FavoriteServiceCargo(FavoriteRepositoryCargo favoriteRepositoryCargo, CargoRepository cargoRepository, UserRepository userRepository) {
         this.favoriteRepositoryCargo = favoriteRepositoryCargo;
         this.cargoRepository = cargoRepository;
+        this.userRepository = userRepository;
     }
 
 
-    public FavoritesCargo addCargo(int userId, int cargoId) {
+    public boolean addCargo(int userId, int cargoId) {
+        Cargo ftDBEntity = cargoRepository.findById(cargoId).orElse(null);
+        User userDBEntity = userRepository.findById(userId).orElse(null);
+        if (ftDBEntity == null || userDBEntity == null)
+            return false;
         FavoritesCargo btf = new FavoritesCargo();
         btf.setUserId(userId);
         btf.setCargoId(cargoId);
-        return favoriteRepositoryCargo.saveAndFlush(btf);
+        favoriteRepositoryCargo.saveAndFlush(btf);
+        return true;
     }
 
     public ArrayList<FavoritesCargo> findAllByUserId(int id) {
         return favoriteRepositoryCargo.findByUserId(id);
     }
-
-    public void deleteById(int id) {
+    @Transactional
+    public boolean deleteById(int id) {
+        FavoritesCargo ftDBEntity = favoriteRepositoryCargo.findById(id).orElse(null);
+        if (ftDBEntity == null)
+            return false;
         favoriteRepositoryCargo.deleteById(id);
+        return true;
     }
 
     public FavoritesCargo getCargoById(int id) {
         return favoriteRepositoryCargo.findById(id).orElse(null);
+    }
+
+    public boolean deleteCurrentUserCargo(Integer cargoId, Integer userId) {
+        FavoritesCargo fcDBEntity = favoriteRepositoryCargo.findById(cargoId).orElse(null);
+        if (fcDBEntity == null)
+            return false;
+        favoriteRepositoryCargo.deleteById(cargoId);
+        favoriteRepositoryCargo.flush();
+        return true;
+    }
+
+    public boolean addCurrentUserCargo(int userId, int cargoId) {
+        Cargo ftDBEntity = cargoRepository.findById(cargoId).orElse(null);
+        if (ftDBEntity == null)
+            return false;
+        FavoritesCargo btf = new FavoritesCargo();
+        btf.setUserId(userId);
+        btf.setCargoId(cargoId);
+        favoriteRepositoryCargo.saveAndFlush(btf);
+        return true;
     }
 }
