+++ src/main/java/com/tms/service/FavoriteServiceTransport.java
@@ -1,7 +1,9 @@
 package com.tms.service;
 
+import com.tms.domain.Cargo;
 import com.tms.domain.FavoritesTransport;
 import com.tms.domain.Transport;
+import com.tms.domain.User;
 import com.tms.repository.CargoRepository;
 import com.tms.repository.FavoriteRepositoryTransport;
 import com.tms.repository.TransportRepository;
@@ -18,26 +20,35 @@ public class FavoriteServiceTransport {
 
     FavoriteRepositoryTransport favoriteRepositoryTransport;
 
-   TransportRepository transportRepository;
-   UserRepository userRepository;
+    TransportRepository transportRepository;
+    UserRepository userRepository;
 
     @Autowired
-    public FavoriteServiceTransport(FavoriteRepositoryTransport favoriteRepositoryTransport, CargoRepository cargoRepository, TransportRepository transportRepository,UserRepository userRepository) {
+    public FavoriteServiceTransport(FavoriteRepositoryTransport favoriteRepositoryTransport, CargoRepository cargoRepository, TransportRepository transportRepository, UserRepository userRepository) {
         this.favoriteRepositoryTransport = favoriteRepositoryTransport;
         this.transportRepository = transportRepository;
         this.userRepository = userRepository;
     }
-    public FavoritesTransport addTransport(int userId, int transportId) {
-        FavoritesTransport btf = new FavoritesTransport();
-            btf.setUserId(userId);
-            btf.setTransportId(transportId);
-            return favoriteRepositoryTransport.saveAndFlush(btf);
 
+    public boolean addTransport(int userId, int transportId) {
+        Transport ftDBEntity = transportRepository.findById(transportId).orElse(null);
+        User userDBEntity = userRepository.findById(userId).orElse(null);
+        if (ftDBEntity == null || userDBEntity==null )
+            return false;
+        FavoritesTransport btf = new FavoritesTransport();
+        btf.setUserId(userId);
+        btf.setTransportId(transportId);
+        favoriteRepositoryTransport.saveAndFlush(btf);
+        return true;
     }
 
     @Transactional
-    public void deleteById(int id) {
+    public boolean deleteById(int id) {
+        FavoritesTransport ftDBEntity = favoriteRepositoryTransport.findById(id).orElse(null);
+        if (ftDBEntity == null)
+            return false;
         favoriteRepositoryTransport.deleteById(id);
+        return true;
     }
 
     public ArrayList<FavoritesTransport> findAllByUserId(int id) {
@@ -46,6 +57,28 @@ public class FavoriteServiceTransport {
 
     public FavoritesTransport getTransportById(int id) {
         return favoriteRepositoryTransport.findById(id).orElse(null);
+    }
+
+
+    public boolean deleteCurrentUserTransport(Integer transportId, Integer userId) {
+        FavoritesTransport ftDBEntity = favoriteRepositoryTransport.findById(transportId).orElse(null);
+        if (ftDBEntity == null || ftDBEntity.getUserId() != userId)
+            return false;
+        favoriteRepositoryTransport.deleteById(transportId);
+        favoriteRepositoryTransport.flush();
+        return true;
+    }
+
+    public boolean addCurrentTransport(int userId, int transportId) {
+        Transport ftDBEntity = transportRepository.findById(transportId).orElse(null);
+        if (ftDBEntity == null)
+            return false;
+        FavoritesTransport btf = new FavoritesTransport();
+        btf.setUserId(userId);
+        btf.setTransportId(transportId);
+        favoriteRepositoryTransport.saveAndFlush(btf);
+        return true;
+
     }
 
 }
